requirements.txt
anyio @ file:///D:/Doston/Projects/IntellectApi/moduls/anyio-3.6.2-py3-none-any.whl
asyncpg @ file:///D:/Doston/Projects/IntellectApi/moduls/asyncpg-0.26.0-cp39-cp39-win_amd64.whl
bcrypt @ file:///D:/Doston/Projects/IntellectApi/moduls/bcrypt-4.0.1-cp36-abi3-win_amd64.whl
certifi @ file:///D:/Doston/Projects/IntellectApi/moduls/certifi-2022.9.24-py3-none-any.whl
charset-normalizer @ file:///D:/Doston/Projects/IntellectApi/moduls/charset_normalizer-2.0.12-py3-none-any.whl
click @ file:///D:/Doston/Projects/IntellectApi/moduls/click-8.1.3-py3-none-any.whl
colorama @ file:///D:/Doston/Projects/IntellectApi/moduls/colorama-0.4.6-py2.py3-none-any.whl
cx-Oracle @ file:///D:/Doston/Projects/IntellectApi/cx_Oracle-8.3.0-cp39-cp39-win_amd64.whl
ecdsa @ file:///D:/Doston/Projects/IntellectApi/moduls/ecdsa-0.18.0-py2.py3-none-any.whl
fastapi @ file:///D:/Doston/Projects/IntellectApi/moduls/fastapi-0.85.1-py3-none-any.whl
greenlet @ file:///D:/Doston/Projects/IntellectApi/moduls/greenlet-1.1.3.post0-cp39-cp39-win_amd64.whl
h11 @ file:///D:/Doston/Projects/IntellectApi/moduls/h11-0.14.0-py3-none-any.whl
idna @ file:///D:/Doston/Projects/IntellectApi/moduls/idna-3.4-py3-none-any.whl
passlib @ file:///D:/Doston/Projects/IntellectApi/moduls/passlib-1.7.4-py2.py3-none-any.whl
psycopg2 @ file:///D:/Doston/Projects/IntellectApi/moduls/psycopg2-2.9.5-cp39-cp39-win_amd64.whl
psycopg2-binary @ file:///D:/Doston/Projects/IntellectApi/moduls/psycopg2_binary-2.9.5-cp39-cp39-win_amd64.whl
pyasn1 @ file:///D:/Doston/Projects/IntellectApi/moduls/pyasn1-0.4.8-py2.py3-none-any.whl
pydantic @ file:///D:/Doston/Projects/IntellectApi/moduls/pydantic-1.10.2-cp39-cp39-win_amd64.whl
python-dotenv @ file:///D:/Doston/Projects/IntellectApi/moduls/python_dotenv-0.21.0-py3-none-any.whl
python-jose @ file:///D:/Doston/Projects/IntellectApi/moduls/python_jose-3.3.0-py2.py3-none-any.whl
python-multipart==0.0.5
requests @ file:///D:/Doston/Projects/IntellectApi/moduls/requests-2.28.1-py3-none-any.whl
rsa @ file:///D:/Doston/Projects/IntellectApi/moduls/rsa-4.9-py3-none-any.whl
six @ file:///D:/Doston/Projects/IntellectApi/moduls/six-1.16.0-py2.py3-none-any.whl
sniffio @ file:///D:/Doston/Projects/IntellectApi/moduls/sniffio-1.3.0-py3-none-any.whl
SQLAlchemy @ file:///D:/Doston/Projects/IntellectApi/moduls/SQLAlchemy-1.4.42-cp39-cp39-win_amd64.whl
SQLAlchemy-Utils @ file:///D:/Doston/Projects/IntellectApi/moduls/SQLAlchemy_Utils-0.38.3-py3-none-any.whl
starlette @ file:///D:/Doston/Projects/IntellectApi/moduls/starlette-0.20.4-py3-none-any.whl
typing_extensions @ file:///D:/Doston/Projects/IntellectApi/moduls/typing_extensions-4.4.0-py3-none-any.whl
urllib3 @ file:///D:/Doston/Projects/IntellectApi/moduls/urllib3-1.26.12-py2.py3-none-any.whl
uvicorn @ file:///D:/Doston/Projects/IntellectApi/moduls/uvicorn-0.19.0-py3-none-any.whl
